requirements-260107.txt
azure-ai-agentserver-agentframework==1.0.0b10
agent-framework==1.0.0b260107
agent-framework-a2a==1.0.0b260107
agent-framework-ag-ui==1.0.0b260107
agent-framework-anthropic==1.0.0b260107
agent-framework-azure-ai==1.0.0b260107
agent-framework-azure-ai-search==1.0.0b260107
agent-framework-azurefunctions==1.0.0b260107
agent-framework-chatkit==1.0.0b260107
agent-framework-copilotstudio==1.0.0b260107
agent-framework-core==1.0.0b260107
agent-framework-declarative==1.0.0b260107
agent-framework-devui==1.0.0b260107
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260107
agent-framework-ollama==1.0.0b260107
agent-framework-purview==1.0.0b260107
agent-framework-redis==1.0.0b260107
azure-monitor-opentelemetry
debugpy
--pre
agent-dev-cli